// src/shared/client-services/model/memberDto.ts
/**
 * SkinHubApp API
 * A Skin Cream, Soap and Beauty Analysis Forum
 *
 * OpenAPI spec version: v1
 * Contact: 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface MemberDto {
    id?: number;
    username?: string;
    emailAddress?: string;
    firstname?: string;
    middlename?: string;
    lastname?: string;
    color?: MemberDto.ColorEnum;
    gender?: MemberDto.GenderEnum;
    dateOfBirth?: Date;
    age?: number;
}
export namespace MemberDto {
    export type GenderEnum = 1 | 2 | 3;
    export const GenderEnum = {
        NUMBER_1: 1 as GenderEnum,
        NUMBER_2: 2 as GenderEnum,
        NUMBER_3: 3 as GenderEnum
    }

    export type ColorEnum = 1 | 2 | 3 | 4;
    export const ColorEnum = {
      NUMBER_1: 1 as ColorEnum,
      NUMBER_2: 2 as ColorEnum,
      NUMBER_3: 3 as ColorEnum,
      NUMBER_4: 4 as ColorEnum
    }
}